fix(test): extract pure reaper module; robust dead-PID in unit test
Review: the unit test imported embedded-postgres-backend, whose top-level
'import EmbeddedPostgres from embedded-postgres' made the no-DB unit suite fail
to load (Cannot find package embedded-postgres). Extract the pure reaper
(reapStaleClustersIn / clusterIsLive / STALE_CLUSTER_MS) into its own module
(node:fs/path only); the backend re-imports it, and the unit test imports the
pure module — no native dep pulled.

Also use a spawn-and-reaped PID for the dead-owner case instead of a magic
out-of-range number, which process.kill(pid,0) handles inconsistently across
runtimes. Validated under node: dead reaped, live kept, fresh kept.
